Refactor: 계산기의 동작을 제어하는 클래스 추상화
